Add support for emitting streamed updates in JSON format
There are scenarios where users may want to have the more or less raw
JSON encoded data available to them, especially when it comes to
ticker event data, which may be processed later on.
We do not really support intercepting this data at the "raw" request
level, but we can support re-encoding the received entity. This change
enables exactly this support. It introduces a new option, -j/--json,
that can be used to enable this support.