Fail gracefully on ticker event parsing errors
For some reason we were using Result::unwrap on the result representing
the JSON deserialization of ticker events. We have seen this unwrap
cause a panic to the program due to a known flaw in the polygon crate.
Regardless of the cause, we should not be so prone to packing and so
with this change we replace the unwrap with a more graceful reporting of
the error.