Set type_length_limit
With arbitrary code changes and/or changes to the tool chain we see
build breakages caused by some compiler internal type length limit being
exceeded. So far this has only been the case in other crates, but now
this application is also affected when compiling in release mode.
Because there is not much we can do other than bumping the limit every
time we see such a complaint, this change sets a rather high initial
value in the hope that this will be good enough for a while.